backend/AI_Ejercicios/requirements.txt
fastapi
uvicorn[standard]
pydantic
motor
python-dotenv
pymongo
azure-ai-inference
httpx
